FIxed the issue of getting the namespace URIs for the Schema element. Added the programatic functionality required for the generation of the Message Receiver. The template is not included.

+++ modules/core/src/org/apache/axis/context/OperationContext.java
@@ -144,7 +144,7 @@ public class OperationContext extends AbstractContext {
      * @return
      * @throws AxisFault
      */
-    public MessageContext getMessageContext(byte messageLabel) throws AxisFault {
+    public MessageContext getMessageContext(int messageLabel) throws AxisFault {
         if (messageLabel == WSDLConstants.MESSAGE_LABEL_IN) {
             return inMessageContext;
         } else if (messageLabel == WSDLConstants.MESSAGE_LABEL_OUT) {
